\name{merpsychophysics-package}
\alias{merpsychophysics-package}
\alias{merpsychophysics}
\docType{package}
\title{
\packageTitle{merpsychophysics}
}
\description{
\packageDescription{merpsychophysics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{merpsychophysics}
\packageIndices{merpsychophysics}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{merpsychophysics}

Maintainer: \packageMaintainer{merpsychophysics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
