% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_utils.R
\name{MixDelta}
\alias{MixDelta}
\title{PSE/JND for Multivariable GLMM with Delta Methods}
\usage{
MixDelta(xplode.obj, alpha = 0.05, p = 0.75)
}
\arguments{
\item{xplode.obj}{an object of class \code{xplode.obj}. The fitted model
(object of class \code{\link[lme4]{merMod}}) from \code{xplode.obj} includes
one continuous predictor and one factorial predictor.}

\item{alpha}{significance level of the confidence intervals. Default is 0.05.}

\item{p}{probability for JND (default 75\%)}
}
\value{
A list, whose lenght is equal to the levels of the factorial predictor, i.
Each cell of the list is equal to the output of \code{delta.psy.probit} applied to
a multivariable model whose baseline is level i of the factorial predictor.
}
\description{
Estimate the Point of Subjective Equivalence (PSE), the Just Noticeable
Difference (JND) and the related Standard Errors for a multivariate distribution
(i.e. one continuous and one factorial predictor) by means of 
delta method.
The method applies to multivariable GLMM having a \emph{probit} link function.
The function is based on a recursive use of \code{\link[lme4]{glmer}} and
\code{\link{MixDelta}}
}
\details{
The function \code{MixDelta} is based on a recursive use of
\code{glmer} and \code{PsychDelta} to multivariable GLMM including
continuous and factorial predictors. The same caveats of \code{PsychDelta}
apply (e.g., confidence interval based on normality assumption).
}
\examples{
library(lme4)
data(vibro_exp3)
formula.mod <- cbind(faster, slower) ~ speed * vibration + (1 + speed| subject)
mod <- glmer(formula = formula.mod, family = binomial(link = "probit"), data = vibro_exp3)
xplode.mod <- xplode(model = mod, name.cont = "speed", name.factor = "vibration")
MixDelta(xplode.mod)

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26
}
\seealso{
\code{\link[lme4]{glmer}} for Generalized Linear Mixed Models. 
\code{\link{MixDelta}} for PSE and JND estimation from a univariable GLMM using delta method. 
\code{\link{pseMer}} for bootstrap-based confidence intervals of psychometric parameters.
}
\keyword{DeltaMethod}
\keyword{GLMM}
\keyword{Multivariable}
