% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_utils.R
\name{MixPlot}
\alias{MixPlot}
\title{Plotting Psychometric Functions given GLMM}
\usage{
MixPlot(xplode.obj, showData = TRUE)
}
\arguments{
\item{xplode.obj}{an object of class \code{\link{xplode}}}

\item{showData}{}
}
\description{
Plot binomial data and the fitted GLMM from an object of class \code{\link{xplode}}.
}
\details{

}
\examples{
library(lme4)
formula.mod <- cbind(faster, slower) ~ speed + (1 + speed| subject)
mod <- glmer(formula = formula.mod, family = binomial(link = "probit"),
              data = vibro_exp3[vibro_exp3$vibration == 0,])
define.mod <- list(pf1 = list(intercept = 1, slope = 2))
xplode.mod <- xplode(model = mod, name.cont = "speed", define.pf = define.mod)
myplot <- MixPlot_1.0(xplode.mod, pf = 1,  p05line = FALSE, x.ref = 8.5, x.range = c(1,16),
                  col = TRUE, x.label = "Stimulus Speed", y.label = "Predicted Response")
                  

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26
}
\seealso{
\code{\link{xplode}} for objects of class \code{xplode}.
}
