% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseMer.R
\name{pseMer}
\alias{pseMer}
\title{PSE/JND for GLMM Using Bootstrap Methods}
\usage{
pseMer(mer.obj, B = 200, FUN = NULL, alpha = 0.05, ci.type = c("norm",
  "basic", "perc"), beep = F)
}
\arguments{
\item{mer.obj}{An object of class \code{"\linkS4class{merMod}"}.}

\item{B}{integer: the number of bootstrap samples.}

\item{FUN}{An optional, custom made function to specify the required parameters to be estimated.
if NULL, \code{pseMer()} will estimate the PSE and the JND of a univariable GLMM.}

\item{alpha}{Significance level of the confidence interval.}

\item{ci.type}{A vector of character strings representing the type of intervals required. The value 
should be any subset of the values c("norm","basic", "stud", "perc", "bca") or simply "all" which will 
compute all five types of intervals. "perc" should be always included for the summary table.}

\item{beep}{Logical. If TRUE, a "ping" sound alerts that the simulation is complete.}
}
\value{
\code{pseMer} returns a list of length 3 including a summary table (Estimate, Standard Error,
Inferior and Superior Confidence Interval of the parameters) and the output of  \code{\link[lme4]{bootMer}} 
and \code{\link[boot]{boot.ci}} functions, for further analises. Confidence Intervals in the summary table are
based on the percentile method.
}
\description{
Estimates the Point of Subjective Equivalence (PSE), the Just Noticeable
Difference (JND) and the related Standard Errors by means of Bootstrap Method.
}
\details{
\code{pseMer} estimates PSE and JND (and additional user defined paremters) from a 
fitted GLMM model (class \code{"\linkS4class{merMod}"}).
}
\note{
A first custom function was written in 2012 for the non-CRAN package MERpsychophisics,
based on the algorithm in Moscatelli et al. (2012). The current function is a simple wrapper
of \code{lme4::bootMer()} and \code{boot::boot.ci()} functions.

Increasing the nuber of bootstrap samples (\code{B}) makes the estimate more reliable. 
However, this will also increase the duration of the computation.
}
\examples{
library(lme4)
data(vibro_exp3)
formula.mod <- cbind(faster, slower) ~ speed + (1 + speed| subject)
mod <- glmer(formula = formula.mod, family = binomial(link = "probit"), 
              data = vibro_exp3[vibro_exp3$vibration == 0,])
pse.boot <- pseMer(mod, B = 100, ci.type = c("norm", "perc"))
pse.boot[1]

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26

Bates, D., Mächler, M., Bolker, B., & Walker, S. (2015). Fitting Linear Mixed-Effects 
Models Using lme4. Journal of Statistical Software, 67(1), 51. https://doi.org/10.18637/jss.v067.i01
}
\seealso{
\code{\link[lme4]{bootMer}} from \code{lme4} package and \code{\link[boot]{boot.ci}} from \code{boot} package. 
The "ping" sound is provided by \code{\link[beepr]{beep}} function from the \code{beepr} package.
}
