% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychBoot}
\alias{PsychBoot}
\title{Internal function: PSE/JND from GLM or GNLM with Bootstrapping}
\usage{
PsychBoot(model_obj, data, response, stimuli, alpha, p, R = 100)
}
\arguments{
\item{model_obj}{the fitted model. An object of class \code{gnlm}.}

\item{data}{A data frame containing the variables used in the model.}

\item{response}{A character vector with the name of the response variables. It is used to create the matrix of the binomial response. It must be provided for gnlm.}

\item{stimuli}{A character string with the name of the stimulus variable. It must be provided for gnlm.}

\item{alpha}{significance level of the confidence interval.Default is 0.05 (95\% confidence interval).}

\item{p}{probability value relative to the JND upper limit. Default is 0.75 (value for 50\% JND).}

\item{R}{Integer. Number of bootstrap samples.}
}
\description{
Internal function: PSE/JND from GLM or GNLM with Bootstrapping
}
