% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychEstimate}
\alias{PsychEstimate}
\title{PsychEstimate - Estimate Psychological Statistics}
\usage{
PsychEstimate(
  model_obj,
  alpha = 0.05,
  p = 0.75,
  method,
  data = NULL,
  response = NULL,
  stimuli = NULL
)
}
\arguments{
\item{model_obj}{An object of class 'gnlm' or 'glm' representing the fitted model.}

\item{alpha}{significance level of the confidence interval.Default is 0.05 (95\% confidence interval).}

\item{p}{probability value relative to the JND upper limit. Default is 0.75 (value for 50\% JND).}

\item{method}{The estimation method. Options are "delta" or "boot".}

\item{data}{A data frame containing the variables used in the model.}

\item{response}{A character vector with the name of the response variables. It is used to create the matrix of the binomial response. It must be provided for gnlm.}

\item{stimuli}{A character string with the name of the stimulus variable. It must be provided for gnlm.}

\item{...}{description}
}
\value{
This function returns estimates of psychometric parameters.
}
\description{
This function returns Point of Subjective Equivalence (PSE), Just Noticeable
Difference (JND), and related Standard Errors of an individual participant using 
either the delta method or bootstrapping.
}
\examples{
data.S1 <- subset(simul_data, Subject == "S1")
model.glm = glm(formula = cbind(Longer, Total - Longer) ~ X,
family = binomial(link = "probit"), data = data.S1)
PsychEstimate(model.glm, method = "delta")

}
