% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_glm}
\alias{PsychFunction_glm}
\title{Internal Function: Fit Generalized Linear Models}
\usage{
PsychFunction_glm(formula, response, stimuli, model, link, data)
}
\arguments{
\item{formula}{A formula specifying the model. If NULL, 'response' and 'stimuli' must be provided.}

\item{response}{A character vector of response variables.}

\item{stimuli}{A character string of the stimulus variable.}

\item{model}{A character string specifying the model ('glm' or 'brglm').}

\item{link}{A character string specifying the link function (possible options: 'probit', 'logit', 'weibull').}

\item{data}{A data frame containing the variables specified in the formula (or in the response and stimuli arguments).}
}
\value{
A list containing:
\item{model}{The fitted model.}
\item{flag}{A logical value indicating whether the a Bias Reduced GLM (brglm) is recommended.}
}
\description{
This function fits a generalized linear model for psychometric functions with glm() or brglm().
}
\seealso{
[\code{\link{PsychFunction}} ]
}
