% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_gnlm}
\alias{PsychFunction_gnlm}
\title{Internal Function: Fit Generalized Nonlinear Models}
\usage{
PsychFunction_gnlm(model_glm, response, stimuli, link, data, guess, lapse)
}
\arguments{
\item{model_glm}{A glm object obtained from PsychFunction_glm. This is used to extract starting estimate of the parameters of the non-linear model.}

\item{response}{A character vector with the name of the response variables.}

\item{stimuli}{A string with name of the stimuli variable.}

\item{link}{A string specifying the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{data}{A data frame containing the variables.}

\item{guess, lapse}{Logical or numeric values indicating whether to include guessing and lapse parameters, respectively. Default is FALSE for both. If parameters are FALSE, they are not included in the model. If TRUE, parameters are estimated with a randomly assigned starting value. If numeric, the value is used as a starting estimate.}
}
\value{
A list containing:
\item{model_gnlm}{The fitted generalized non-linear model.}
\item{gnlr_coeff}{The estimated coefficients (see notes). }
}
\description{
This function fits a generalized nonlinear model for psychometric functions.
}
\note{
For easier use of the \code{\link{gnlr}} function, the vector of stimuli is defined as a global variable. It's recommended to avoid global variables for better code maintainability and to prevent potential conflicts. This implementation will be changed in future releases.
When guess and lapse parameters are present, the fitted coefficients returned by the \code{glnr} function undergo a log-transformation. This transformation is applied as a result of fitting the parameters of the model using the exponential function, which ensures that the coefficients remain positive.
}
\seealso{
[\code{\link{PsychFunction}}] [\code{\link{switch_mu_function}} ]
}
