% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils_multi.R
\name{PsychParameters}
\alias{PsychParameters}
\title{Fit multiple psychometric functions with GLM}
\usage{
PsychParameters(model_list, se = TRUE)
}
\description{
Fit multiple psychometric functions with GLM
}
