% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setGlobalVar}
\alias{setGlobalVar}
\title{Internal Function: Set global variable}
\usage{
setGlobalVar(data, stimuli)
}
\arguments{
\item{data}{dataset}

\item{stimuli}{A string of the stumuli variable.}
}
\description{
Setting a global variable is necessary in order to correctly define the function to fit with gnlm.
}
\seealso{
[PsychFunction_gnlm()]
}
