% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{StartEstimate}
\alias{StartEstimate}
\title{Internal Function: Find start estimate parameters for gnlm}
\usage{
StartEstimate(model_glm, link, guess, lapse)
}
\arguments{
\item{model_glm}{A glm object obtained from PsychFunction_glm. This is used to extract starting estimate of the parameters of the non-linear model.}

\item{link}{A character string specifying the link function. For generalized linear models, it defines the link function specified in the family object. See \code{\link[stats]{family}} for details of family functions for the binomial family. For generalized non-linear models, it defines the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{guess, lapse}{Logical or numeric values indicating whether to include guessing and lapse parameters, respectively. These parameters are only used if \code{model = 'gnlm'}. Default is FALSE for both. If parameters are FALSE, they are not included in the model. If TRUE, parameters are estimated with a randomly assigned starting value. If numeric, the value is used as a starting estimate.}
}
\description{
Internal Function: Find start estimate parameters for gnlm
}
