% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{switch_mu_function}
\alias{switch_mu_function}
\title{Internal Function: Switch mu Function}
\usage{
<<<<<<< HEAD
switch_mu_function(x_values, func_name, data, stimuli, gamma, lambda)
=======
switch_mu_function(link, gamma, lambda)
>>>>>>> 68028e7c9192c5fecae41ef5fb1d74a14c12d29f
}
\arguments{
\item{link}{A string specifying the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{gamma, lambda}{parameters indicating whether to include guessing and lapse parameters, respectively. If parameters are FALSE, they are not included in the model. If numeric, they are the starting estimates used in \code{\link{gnlr}}.}
}
\value{
A function representing the selected mu function.
}
\description{
This function switches between different mu functions based on the provided parameters.
}
\note{
Instead of directly fitting guess and lapse parameters, this function uses \code{exp(p[3])} and \code{exp(p[4])}. By doing so, we force the estimations to be non-negative.
}
\seealso{
[\code{\link{PsychFunction}}] [\code{\link{PsychFunction_gnlm}}] [\code{\link{gnlr}}]
}
