% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{x_values}
\alias{x_values}
\title{My Global Variable}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
x_values
}
\description{
Description of my_global_var.
}
\keyword{datasets}
