% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_utils.R
\name{MixPlot}
\alias{MixPlot}
\title{Plotting Psychometric Functions given GLMM}
\usage{
MixPlot(xplode.obj, showData = TRUE)
}
\arguments{
\item{showData}{}

\item{xplode}{}
}
\description{
Plot a psychometric function given an object of class \code{\link[stats]{glm}} or \code{\link[brglm]{brglm}}.
}
\details{

}
\examples{

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26

Knoblauch, K., & Maloney, L. T. (2012). Modeling psychophysical data in R (Vol. 32). 
Springer Science & Business Media.
}
\seealso{
\code{\link[stats]{glm}} for for Generalized Linear Models.
\code{\link{PsychFunction}} for estimation of PSE and JND.
}
