% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_utils.R
\name{MixPlot}
\alias{MixPlot}
\title{Plotting Psychometric Functions given GLMM}
\usage{
MixPlot(xplode.obj, facet_by = NULL, showData = TRUE)
}
\arguments{
\item{xplode.obj}{an object of class \code{\link{xplode}}}

\item{facet_by}{}

\item{showData}{}
}
\description{
Plot binomial data and the fitted GLMM from an object of class \code{\link{xplode}}.
}
\details{

}
\examples{
library(lme4)
formula.mod <- cbind(faster, slower) ~ speed *vibration + (1 + speed| subject)
mod <- glmer(formula = formula.mod, family = binomial(link = "probit"),
              data = vibro_exp3)
xplode.mod <- xplode(model = mod, name.cont = "speed", name.factor = "vibration")
myplot <- MixPlot(xplode.mod)
                  

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26
}
\seealso{
\code{\link{xplode}} for objects of class \code{xplode}.
}
