% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_utils.R
\name{pseMer}
\alias{pseMer}
\title{PSE/JND from GLMM Estimates Using Bootstrap Method}
\usage{
pseMer(
  mer.obj,
  B = 200,
  FUN = NULL,
  alpha = 0.05,
  ci.type = c("norm", "basic", "perc"),
  beep = F
)
}
\arguments{
\item{mer.obj}{an object of class \code{\linkS4class{merMod}}.}

\item{B}{integer. Number of bootstrap samples.}

\item{FUN}{an optional, custom made function to specify the required parameters to be estimated.
if NULL, \code{pseMer()} estimates PSE and JND of a univariable GLMM with a single intercept and slope.}

\item{alpha}{significance level of the confidence intervals. Default is 0.05 (value for 95\% confidence interval).}

\item{ci.type}{vector of character strings representing the type of intervals required. The value 
should be any subset of the values c("norm","basic", "stud", "perc", "bca") or simply "all" which will 
compute all five types of intervals. "perc" should be always included for the summary table.}

\item{beep}{logical. If TRUE, a "ping" sound alerts that the simulation is complete. Default is FALSE.}
}
\value{
\code{pseMer} returns a list of length 3 including a summary table (Estimate, Standard Error,
Inferior and Superior Confidence Interval of the parameters) and the output of  \code{\link[lme4]{bootMer}} 
and \code{\link[boot]{boot.ci}} functions, for further analyses. Confidence Intervals in the summary table are
based on the percentile method.
}
\description{
Estimates the Point of Subjective Equivalence (PSE), the Just Noticeable
Difference (JND) and the related Standard Errors by means of Bootstrap Method, 
given an object of class \code{\link[lme4]{merMod}}.
}
\details{
\code{pseMer} estimates PSE and JND (and additional user defined parameters) from a 
fitted GLMM model (class \code{"\linkS4class{merMod}"}).
}
\note{
A first custom function was written in 2012 for the non-CRAN package MERpsychophisics,
based on the algorithm in Moscatelli et al. (2012). The current function is a simple wrapper
of \code{lme4::bootMer()} and \code{boot::boot.ci()} functions.

Increasing the number of bootstrap samples (\code{B}) makes the estimate more reliable. 
However, this will also increase the duration of the computation.
}
\examples{
## Example 1: estimate pse/jnd of a univariable GLMM
library(lme4)
data(vibro_exp3)
formula.mod1 <- cbind(faster, slower) ~ speed + (1 + speed| subject)
mod1 <- glmer(formula = formula.mod1, family = binomial(link = "probit"), 
              data = vibro_exp3[vibro_exp3$vibration == 0,])
\dontshow{BootEstim.1a <- pseMer(mod1, B = 5, ci.type = c("perc"))}
\donttest{BootEstim.1 <- pseMer(mod1, B = 100, ci.type = c("perc"))}

## Example 2: specify custom parameters for bootstrap estimation of a 
# multivariate model

formula.mod2 <- cbind(faster, slower) ~ speed * vibration + (1 + speed| subject)
mod2 <- glmer(formula = formula.mod2, family = binomial(link = "probit"), 
               data = vibro_exp3)
              
fun2mod = function(mer.obj){
#allocate space: 4 parameters (jnd_0Hz, jnd_32Hz, pse_0Hz, pse_32Hz) j
jndpse = vector(mode = "numeric", length = 4)
names(jndpse) = c("jnd_0Hz","jnd_32Hz", "pse_0Hz", "pse_32Hz")
jndpse[1] = qnorm(0.75)/fixef(mer.obj)[2] #jnd_0Hz
jndpse[2] = qnorm(0.75)/(fixef(mer.obj)[2] + fixef(mer.obj)[4]) #jnd_32Hz
jndpse[3] = -fixef(mer.obj)[1]/fixef(mer.obj)[2] #pse_0Hz
jndpse[4] = -(fixef(mer.obj)[1] + fixef(mer.obj)[3])/(fixef(mer.obj)[2] 
               + fixef(mer.obj)[4]) #pse_32Hz
return(jndpse)
}

\donttest{BootEstim.2 = pseMer(mod2, B = 100, FUN = fun2mod)}

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26

Bates, D., Mächler, M., Bolker, B., & Walker, S. (2015). Fitting Linear Mixed-Effects 
Models Using lme4. Journal of Statistical Software, 67(1), 51. https://doi.org/10.18637/jss.v067.i01
}
\seealso{
\code{\link[lme4]{bootMer}} from \code{lme4} package and \code{\link[boot]{boot.ci}} from \code{boot} package.
}
\keyword{Bootstrap}
\keyword{GLMM}
