% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction}
\alias{PsychFunction}
\title{Fit Psychometric Function from Single-Subject Response}
\usage{
PsychFunction(
  formula = NULL,
  response = NULL,
  stimuli = NULL,
  model = "glm",
  link = "probit",
  data,
  guess = FALSE,
  lapse = FALSE,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the model (glm or brglm). If NULL, 'response' and 'stimuli' must be provided.}

\item{response}{A character vector with the name of the response variables. It is used to create the matrix of the binomial response. It must be provided for gnlm.}

\item{stimuli}{A character string with the name of the stimulus variable. It must be provided for gnlm.}

\item{model}{A character string specifying the model. Possible options are 'glm' for the generalized linear model, 'brglm' for glm with bias reduction, 'gnlm' for generalized non-linear model. Default is 'glm'.}

\item{link}{A character string specifying the link function. For generalized linear models, it defines the link function specified in the family object. See \code{\link[stats]{family}} for details of family functions (for binomial family). For generalized non-linear models, it defines the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{data}{A data frame containing the variables used in the model.}

\item{guess, lapse}{Logical or numeric value indicating whether to include a guessing parameter in the fitted psychometric function. Only used if model = 'gnlm'. Default is FALSE.}

\item{...}{Additional arguments included for back compatibility. Deprecated arguments from older package versions ('ps.formula', 'ps.link', 'ps.data', 'br') will be automatically handled.}
}
\value{
A list containing:
\item{glm}{The fitted GLM model.}
\item{recomment_br}{A logical value indicating whether bias reduced is recommended.}
\item{brglm}{The fitted Bias Reduced GLM, if model = "brglm".}
\item{gnlm}{The fitted generalized non linear model (gnlm), if model = "gnlm".}
}
\description{
This function provides an interface for fitting psychometric functions using various models: generalized linear model (glm), glm with bias reduction (brglm), generalized non-linear model (gnlm).
}
\note{
If 'model' is specified as "gnlm", a generalized linear model (glm) will be fitted to extract initial estimates for the parameters of the non-linear model.
}
\examples{
\dontrun{
PsychFunction(formula = y ~ x, data = my_data)
}
}
