% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_glm}
\alias{PsychFunction_glm}
\title{Internal Function: Fit Generalized Linear Models}
\usage{
PsychFunction_glm(formula, response, stimuli, model, link, data)
}
\arguments{
\item{formula}{A formula specifying the model. If NULL, 'response' and 'stimuli' must be provided.}

\item{response}{A character vector of response variables.}

\item{stimuli}{A character vector of stimulus variables.}

\item{model}{A character string specifying the model ('glm', 'brglm').}

\item{link}{A character string specifying the link function ('probit', 'logit', 'weibull').}

\item{data}{A data frame containing the variables specified in the formula.}
}
\value{
A list containing the fitted model and additional information.
}
\description{
This function fits a generalized linear model for psychometric functions using glm or brglm.
}
