% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_glm}
\alias{PsychFunction_glm}
\title{Internal Function: Fit Generalized Linear Models}
\usage{
PsychFunction_glm(formula, response, stimuli, model, link, data)
}
\arguments{
\item{formula}{A formula specifying the model. If NULL, 'response' and 'stimuli' must be provided.}

\item{response}{A character vector of response variables.}

\item{stimuli}{A character string of the stimulus variable.}

\item{model}{A character string specifying the model ('glm' or 'brglm').}

\item{link}{A character string specifying the link function. For generalized linear models, it defines the link function specified in the family object. See \code{\link[stats]{family}} for details of family functions for the binomial family. For generalized non-linear models, it defines the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{data}{A data frame containing the variables used in the model.}
}
\value{
A list containing:
\item{model}{The fitted model.}
\item{flag}{A logical value indicating whether the a Bias Reduced GLM (brglm) is recommended.}
}
\description{
This function fits a generalized linear model for psychometric functions with glm() or brglm().
}
\seealso{
[\code{\link{PsychFunction}} ]
}
