% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_gnlm}
\alias{PsychFunction_gnlm}
\title{Internal Function: Fit Generalized Nonlinear Models}
\usage{
PsychFunction_gnlm(
  response,
  stimuli,
  data,
  link = NULL,
  gamma = NULL,
  lambda = NULL,
  pmu = NULL,
  model = NULL
)
}
\arguments{
\item{response}{A character vector with the name of the response variables.}

\item{stimuli}{A string with name of the stimuli variable.}

\item{data}{A data frame containing the variables used in the model.}

\item{link}{A character string specifying the link function. For generalized linear models, it defines the link function specified in the family object. See \code{\link[stats]{family}} for details of family functions for the binomial family. For generalized non-linear models, it defines the function being fitted. Possible options are 'probit' for the cumulative normal distribution, 'logit' for the cumulative logit distribution, 'weibull' for cumulative Weibull distribution.}

\item{gamma}{descr}

\item{lambda}{descr}

\item{pmu}{description}

\item{model}{A character string specifying the model. Possible options are 'glm' for the generalized linear model, 'brglm' for glm with bias reduction, 'gnlm' for generalized non-linear model. Default is 'glm'.}
}
\value{
A list containing:
\item{model_gnlm}{The fitted generalized non-linear model.}
\item{gnlr_coeff}{The estimated coefficients (see notes). }
}
\description{
This function fits a generalized nonlinear model for psychometric functions.
}
\note{
For easier use of the \code{\link{gnlr}} function, the vector of stimuli is defined as a global variable. It's recommended to avoid global variables for better code maintainability and to prevent potential conflicts. This implementation will be changed in future releases.
When guess and lapse parameters are present, the fitted coefficients returned by the \code{glnr} function undergo a log-transformation. This transformation is applied as a result of fitting the parameters of the model using the exponential function, which ensures that the coefficients remain positive.
}
\seealso{
[\code{\link{PsychFunction}}] [\code{\link{switch_mu_function}} ]
}
