% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychFunction_gnlm}
\alias{PsychFunction_gnlm}
\title{Internal Function: Fit Generalized Nonlinear Models}
\usage{
PsychFunction_gnlm(model_glm, link, response, data, stimuli, guess, lapse)
}
\arguments{
\item{model_glm}{A glm object obtained from PsychFunction_glm.}

\item{link}{A character string specifying the link function ('probit', 'logit', 'weibull').}

\item{response}{A character vector of response variables.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{stimuli}{string with name of stimuli variable}

\item{guess}{Logical or numeric value indicating whether to include a guessing parameter.}

\item{lapse}{Logical or numeric value indicating whether to include a lapse parameter.}
}
\value{
A gnlr object representing the fitted model.
}
\description{
This function fits a generalized nonlinear model for psychometric functions.
}
