% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils_multi.R
\name{PsychInterpolate}
\alias{PsychInterpolate}
\title{Fit multiple psychometric functions with GLM}
\usage{
PsychInterpolate(model_list, n_points = 100)
}
\description{
Fit multiple psychometric functions with GLM
}
