% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{PsychPlot}
\alias{PsychPlot}
\title{Plotting Psychometric Functions given GLM}
\usage{
PsychPlot(
  mod,
  addTo = NULL,
  showData = TRUE,
  ps.type = "solid",
  ps.size = 1,
  ps.lab = ""
)
}
\arguments{
\item{mod}{mod}

\item{addTo}{addTo}

\item{showData}{show}

\item{ps.type}{line type.}

\item{ps.size}{line size.}

\item{ps.lab}{label.}
}
\description{
Plot a psychometric function given an object of class \code{\link[stats]{glm}} or \code{\link[brglm]{brglm}}.
}
\details{

}
\examples{

}
\references{
Moscatelli, A., Mezzetti, M., & Lacquaniti, F. (2012). Modeling psychophysical data 
at the population-level: The generalized linear mixed model. 
Journal of Vision, 12(11):26, 1-17. https://doi.org/10.1167/12.11.26

Knoblauch, K., & Maloney, L. T. (2012). Modeling psychophysical data in R (Vol. 32). 
Springer Science & Business Media.
}
\seealso{
\code{\link[stats]{glm}} for for Generalized Linear Models.
\code{\link{PsychFunction}} for estimation of PSE and JND.
}
