% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsySimulate.R
\name{PsySimulate}
\alias{PsySimulate}
\title{Simulate psychophysical data}
\usage{
PsySimulate(fixeff = c(-7, 0.0875), raneff = c(2.4, -0.002, 2e-06),
  nsubjects = 8, pps = 9, ntrials = 40, xint = c(40, 120),
  constant = F)
}
\arguments{
\item{fixeff}{Array of fixed effects. First item is the intercept, second
item is the slope.}

\item{raneff}{Array of random effects. First item is the intercept, second
item is the covariance, third item is the slope.}

\item{nsubjects}{Number of subjects to simulate data for. Default is 8.}

\item{pps}{Number of stimulus levels. Default is 9.}

\item{ntrials}{Number of trials for each stimulus level. Default is 40.}

\item{xint}{Range of the stimulus interval. Default is c(40,120)}

\item{constant}{If set to FALSE, stimulus levels are randomly generated,
uniformly distributed values within the selected interval.
If constant = TRUE, the X interval is divided in  intervals of constant
length. Default is FALSE.}
}
\value{
Simulated dataset
}
\description{
Given the arrays of fixed and random effects, as well as the covariance,
and the characteristic of the simulated experiment (i.e., ) the function
simulates a dataset in which for each subject the following information is
provided: the slope and intercept value of the psychometric function, and
the simulated responses to the stimulus levels that fit that function.
}
\examples{
#simulate dataset (one subject)
datafr.S1 <- MERsimulate(nsubject = 1, constant = T)

#psychometric function to fit single-subject data
plot(Longer/Total ~ X, data = datafr.S1)
fit.S1 = PsychFunction(ps.formula = cbind(Longer, Total - Longer) ~ X
                     ps.link = 'probit', ps.data = datafr.S1,
                     x.range = c(40, 120), ps.lines = T)

fit.S1

PsySimulate()

}
\author{
Alessandro Moscatelli, Priscilla Balestrucci
}
