% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_utils.R
\name{switch_mu_function}
\alias{switch_mu_function}
\title{Internal Function: Switch mu Function}
\usage{
switch_mu_function(func_name, data, stimuli, gamma, lambda)
}
\arguments{
\item{func_name}{A character string specifying the mu function ('probit', 'logit', 'weibull').}

\item{data}{dataset}

\item{stimuli}{name of stimulus variable}

\item{gamma}{A numeric or logical value indicating the gamma parameter.}

\item{lambda}{A numeric or logical value indicating the lambda parameter.}
}
\value{
A function representing the selected mu function.
}
\description{
This function switches between different mu functions based on the provided parameters.
}
